+++ backend/schemas/__init__.py
@@ -1,0 +1,2 @@
+from .sensor import Sensor, SensorCreate
+from .medicao import Medicao, MedicaoCreate
